+++ src/auth/auth.module.ts
@@ -5,8 +5,8 @@ import { JwtModule } from '@nestjs/jwt';
 import { PassportModule } from '@nestjs/passport';
 import { MembersModule } from 'src/members/members.module';
 import { JwtStrategy } from './strategyies/jwt.strategy';
-import { JwtRefreshStrategy } from './strategyies/jwt.refresh.strategy';
-import { JwtOptionalStrategy } from './strategyies/jwt.optional.strategy';
+import { JwtRefreshStrategy } from './strategyies/jwt-refresh.strategy';
+import { JwtOptionalStrategy } from './strategyies/jwt-optional.strategy';
 import { env } from 'src/common/configs/env';
 
 @Module({
