+++ src/auth/guards/jwt-optional.guard.ts
