+++ src/auth/strategyies/jwt-refresh.strategy.ts
@@ -7,7 +7,7 @@ import { PassportStrategy } from '@nestjs/passport';
 import { ExtractJwt, Strategy } from 'passport-jwt';
 import { Request } from 'express';
 import { env } from 'src/common/configs/env';
-import { TokenPayloadDto } from '../../common/dto/tokenPayload.dto';
+import { TokenPayloadDto } from '../../common/dto/token-payload.dto';
 import { MembersService } from 'src/members/members.service';
 interface JwtRefreshPayload extends TokenPayloadDto {
   refreshToken: string;
