+++ src/common/common.module.ts
@@ -1,7 +1,7 @@
 import { Module } from '@nestjs/common';
 import { CommonService } from './common.service';
 import { CommonController } from './common.controller';
-import { PrismaService } from './prismaClient';
+import { PrismaService } from './prisma-client';
 @Module({
   controllers: [CommonController],
   providers: [CommonService, PrismaService],
