+++ src/faq/dto/update-faq.dto.ts
@@ -1,4 +1,4 @@
 import { PartialType } from '@nestjs/mapped-types';
-import { CreateFaqDto } from './create.faq.dto';
+import { CreateFaqDto } from './create-faq.dto';
 
 export class UpdateFaqDto extends PartialType(CreateFaqDto) {}
