+++ src/faq/faq.repository.ts
@@ -1,8 +1,8 @@
 import { Injectable } from '@nestjs/common';
-import { PrismaService } from 'src/common/prismaClient';
-import { CreateFaqDto } from './dto/create.faq.dto';
-import { UpdateFaqDto } from './dto/update.faq.dto';
-import { OrderByQueryDto } from '../common/dto/orderBy.dto';
+import { PrismaService } from 'src/common/prisma-client';
+import { CreateFaqDto } from './dto/create-faq.dto';
+import { UpdateFaqDto } from './dto/update-faq.dto';
+import { OrderByQueryDto } from '../common/dto/order-by.dto';
 @Injectable()
 export class FaqRepository {
   constructor(private readonly prismaClient: PrismaService) {}
